Create tacheServiceimpl et tacheController

+++ TacheService/src/main/java/com/tacheservice/dto/TacheDto.java
@@ -6,8 +6,8 @@ import java.util.Date;
 
 @Data
 public class TacheDto {
-    private Long idTache;
-    private Long idProjet;
+    private int idTache;
+    private int idProjet;
     private String description;
     private Date dateDebut;
     private Date dateFin;
